Switchin required openssl version to 1.1
-- Notes
We have two concurrent problems with openssl:
First, on M1 macs, openssl's libraries are in a different location.
Second -- and more relevant to this particular commit -- eventmachine
is not able to properly use its C extensions with openssl@1.3, which
is what homebrew will install by default. This will fail silently
during the setup/install phase and only become apparent when
attempting to run the application.

The solution is to strictly use openssl@1.1 for the moment.