LG-6562 Preserve personal key after the user cancels personal key confirmation (#6574)
Prior to this commit, when a user cancels personal key entry the personal key is not preserved in the confirmation dialog. This commit makes a change to preserve that value so it remains so the user does not have to re-type.

changelog: Improvements, Personal key confirmation, The "Back" button on the personal key confirmation dialog no longer clears characters previously entered into the form.