Retrieve the VA authorization code
Passed when the VA initiates the inherited proofing process.
This is a WIP; the idea is to simply persist the authorization
code for now.

Create an inherited proofing concern.
- This concern should include all other inherited proofing concerns.
- Test to make sure the auth code is persisted.

Rename request_params and change access modifier.
- VA inherited proofing requirements require query
params appended to the url be accessible.

This commit also renames #request_params to #request_url_params
and changes the access modifier from private to public.