Updating AcuantDocumentState to reflect new sdk changes
-- What
The new 11.7.0 sdk adds a property to the AcuantDocumentState called
`BIG_DOCUMENT`, which is an indication that the image is too close to
the document.

Note also that the enum values have changed as a result.

This is described in the migration doc:
https://github.com/Acuant/JavascriptWebSDKV11/blob/master/docs/MigrationDetails.md

changelog: Improvement, SDK, Updating Acuant SDK to 11.7.0 (LG-6785)