Updating window object checking and fixing tests
-- What
Because of the discrepancy between the 11.5.x and 11.7.x SDKs in how
they declare globals like AcuantJavascriptWebSdk and AcuantCameraUI, I
have now included a check in the acuant.jsx context file that will try
to find those objects on the global window object and, if not present,
simply return the result of using those variables at the global
scope (which is how 11.7.x defines them).

The benefit of this approach is that we don't have to rewrite or edit
a bunch of the frontend tests.

changelog: Improvements, Acuant SDK upgrade, dealing with variable declarations