Updating acuant-camera to tsx and associated typescript changes
-- What
Several problems swirled together in the type system that necessitated
refactoring the acuant-camera component to be a tsx/typescript file.

First, the typechecker could not handle an undefined global
AcuantCameraUI reference, and the JSDoc type annotations seem to not
have a way to deal with this while plain typescript does.

Second, the change to using a guard clause for trying to reason about
the presence of the global variable also threw off the typechecker. In
this commit we've opted for a different approach, which is to simply
log the condition where neither the window nor the global namespace
has a reference to the importan Acuant globals when they are
needed (AcuantJavascriptWebSdk, AcuantCamera, etc). This might not be
the ideal choice, but for the moment it beats fighting with the
typechecker.