LG-6785: Upgrade Acuant SDK to 11.7.0 (#6649)
* Initial commit of files/settings for acuant 11.7.0 sdk

* Updating AcuantDocumentState to reflect new sdk changes

-- What
The new 11.7.0 sdk adds a property to the AcuantDocumentState called
`BIG_DOCUMENT`, which is an indication that the image is too close to
the document.

Note also that the enum values have changed as a result.

This is described in the migration doc:
https://github.com/Acuant/JavascriptWebSDKV11/blob/master/docs/MigrationDetails.md

changelog: Improvement, SDK, Updating Acuant SDK to 11.7.0 (LG-6785)

* Adding Acuant SDK DOM Loading Tests

-- What
There appears to be some difference in how the 11.7 and 11.5 SDKs load
the AcuantJavascriptWebSdk object into the browser's global namespace
once the script tag has been loaded and `loadAcuantSdk` has been
called.

I wrote a s few tests that confirm the behavior I was seeing in my
personal browser interactions and debugging sessions.

changelog: Improvement, SDK, Updating Acuant SDK to 11.7.0 (LG-6785)

* Updating Acuant components to use new globals

-- What
In a previous (11.5.0) version of the SDK, objects like
AcuantJavascriptWebSdk and AcuantCamera were placed onto the browser's
global window object. But in the new (11.7.0) SDK, they are no longer
present on window -- instead, they reside alone in the global
namespace via `let` declarations at the highest scope.

This commit represents elementary changes to two files that reference
these objects in the global namspace, rather than on window where they
don't exist anymore (and were therefore erroring).

The type annotations still need to be dealt with
appropriately. However, the acuant sdk now loads and appears to work
properly in the local manual testing environment.

changelog: Improvements, SDK update, Updating the Acuant SDK

* Switching acuant context file to typescript

* Adding French and Spanish localizations for BIG_DOCUMENT

-- What
The new Acuant SDK image warning BIG_DOCUMENT, which specifies that
the camera is "too close" to the document, has messages that have now
been localized into French and Spanish, as well as the original English.

* Updating window object checking and fixing tests

-- What
Because of the discrepancy between the 11.5.x and 11.7.x SDKs in how
they declare globals like AcuantJavascriptWebSdk and AcuantCameraUI, I
have now included a check in the acuant.jsx context file that will try
to find those objects on the global window object and, if not present,
simply return the result of using those variables at the global
scope (which is how 11.7.x defines them).

The benefit of this approach is that we don't have to rewrite or edit
a bunch of the frontend tests.

changelog: Improvements, Acuant SDK upgrade, dealing with variable declarations

* Fixing AcuantCamera and AcuantCameraUI global checks

-- What
We were checking for the AcuantCamera and AcuantCameraUI objects at
the wrong moment in the flow, which was causing errors when attempting
to load the actual camera. This is now resolved.

changelog: Improvements, SDK upgrade, fixing remaining global variable issues

* Updating comment formatting and adding guards for globals

-- What
 - Switching property and function description comments to use
   docblock format, which is better for VSCode users;
 - Adding guards to global Acuant SDK variable getter methods
 - Removing leftover type annotations from acuant.tsx file

changelog: Improvements, Formatting, Acuant SDK upgrade comment
formatting and guards

* Fixing typo in localization yaml config

* Update app/javascript/packages/document-capture/components/acuant-camera.jsx

Co-authored-by: Andrew Duthie <andrew.duthie@gsa.gov>

* Removing tests for 11.5.0 Acuant SDK loading

* Excluding acuant sdk directories from code-climate checks

* Removing symlinked fr dir in public folder

* Fixing caught eslint errors

-- Notes
I have included an ignore line in acuant-camera.jsx where we refer to
the global AcuantCameraUI without first declaring it. This is due to
the new behavior Acuant introduced in the 11.7.0 SDK, where they are
declaring variables in the global scope (but not in the window)

* Updating acuant-camera to tsx and associated typescript changes

-- What
Several problems swirled together in the type system that necessitated
refactoring the acuant-camera component to be a tsx/typescript file.

First, the typechecker could not handle an undefined global
AcuantCameraUI reference, and the JSDoc type annotations seem to not
have a way to deal with this while plain typescript does.

Second, the change to using a guard clause for trying to reason about
the presence of the global variable also threw off the typechecker. In
this commit we've opted for a different approach, which is to simply
log the condition where neither the window nor the global namespace
has a reference to the importan Acuant globals when they are
needed (AcuantJavascriptWebSdk, AcuantCamera, etc). This might not be
the ideal choice, but for the moment it beats fighting with the
typechecker.

* Adding normalized yaml localization files

* Update app/javascript/packages/document-capture/components/acuant-camera.tsx

Co-authored-by: Andrew Duthie <andrew.duthie@gsa.gov>

* Updating definition of exports in tsx files

Co-authored-by: Andrew Duthie <andrew.duthie@gsa.gov>
Co-authored-by: Eric Gade <ecgade@macbook-m1.lan>