Updating types to deal with new native camera attempts
-- What
With the addition of new checks for `maxAttemptsBeforeNativeCamera`,
which triggers the use of the native camera after a certain number of
failed Acuant attempts, we needed to update some of the type
definitions.

In addition to doing so, we have converted several of the affected
files to full TypeScript (from JSDoc annotations).