Changes to onboarding process (#6560)
* Install mailcatcher via Gemfile

* Step 4 runs `rake db:create` so Step 3 is redundant

* Updating README with macOS openssl instructions

-- Notes
For mailcatcher to install correctly on macOS (M1 and Intel), we need
to install it via brew and link to the correct header location. The
instructions specify how to do that before installing.

* Target openssl v.3

* Restore manual install of mailcatcher for now

include note for openssl issue we resolved

* Switchin required openssl version to 1.1

-- Notes
We have two concurrent problems with openssl:
First, on M1 macs, openssl's libraries are in a different location.
Second -- and more relevant to this particular commit -- eventmachine
is not able to properly use its C extensions with openssl@1.3, which
is what homebrew will install by default. This will fail silently
during the setup/install phase and only become apparent when
attempting to run the application.

The solution is to strictly use openssl@1.1 for the moment.

* Ensuring we build js before testing and updating README

-- What
1. `yarn build` was not being called before tests in the Makefile. So
if you pulled down the repository, ran `make setup` and followed all
the instructions in the readme, tests would still fail because at no
point in that process do js bundles get built. I have added this to
the relevant Makefile actions
2. The README had a typo about openssl1.3 when we require 1.1

changelog: Improvements, Documentation, Updating onboarding
documentation (LG-6972)

Co-authored-by: darth-cheney <eric.gade@gmail.com>