Updating FailedCaptureContext with new dynamic forceNativeCamera
-- What
For clarity in the FailedCaptureContext interface, we've added a
property called `forceNativeCamera` which is a boolean value that will
update in the FailedCaptureContextProvider whenever the
maxAttemptsBeforeNativeCamera is equal to or exceeds the current
failedCaptures value.

Additionally, we finally have some working tests for this. For the
moment, we only get results by forcing maxAttemptsBeforeNativeCamera
to be 0, then clicking the file input. This seems to do the trick.

But we need to figure out how to test a couple of other scenarios, and
how to force failures from that point in a test case.