Updating rails form hints to be labels
-- What
By updating the simple_form initializer wrappers and the configuration
of the output in the validated_form template, we ensure that the hints
-- if present -- are caught by screen readers. We do this by making
the hints a <label> instead of a <div>, and pointing that label to the
same input element as the main form's label. This way both the main
label and the hint are read by screen readers.