Updating forgot password link preamble for screen reader
-- What
On the Re-Enter Password screen for GPO verify, the text "Forgot
password?" that precedes the password reset link is not readable by
screen readers. To deal with this, we add the localized text of
"Forgot password?" to the `title=` attribute of the link itself inside
of the corresponding React component. This ensures that screen readers
will read both the link text and the forgot password hint.

changelog: Improvements, Accessibility, updating password hint