Use full path to load .ruby-version in Gemfile
Some tools (in my case it's the VS Code Ruby extension) attempt to load the `Gemfile` from directories other than the project root. When this happens, `Gemfile` can't load `.ruby-version`, since it's not in the directory it's expecting. This leads to errors from tools complaining about missing `.ruby-version` files. (This is obviously not the _correct_ behavior on the part of the tools, but nevertheless it is _the_ behavior.)
T
This commit updates the `Gemfile` to always explicitly load `.ruby-version` from the same directory `Gemfile` is in.

[skip changelog]