Create separate events for when a letter is requested and when it is enqueued
The letter is not necessarily sent when it is requested. At the end of proofing the user requests a letter, but the letter is not sent until the profile is created after the user enters their password. This commit adds different events for requested and enqueueing the letter so we can track both