Updating corresponding template and localizations / tests
-- What
Because the FSMv2 was still enabled for the React component locally, I
was not seeing the template version. This commit updates the template
and also removes now unused localizations from the config.

I am also updating a couple of Ruby component tests that assumed a
single aria-describedby value on labelled validation components with
hints, whereas now we supply two values to that attribute