Switching field error aria description to come before field hint
-- What
Previously, the order of the aria-describedby attribute values was
reversed from what we would probably want, meaning that if a
validation error was present, the error would be read after the
hint. We want the error to be read first so users know that something
was not right, then provide the hint.