LG-7073 Hint text on the form fields for screen readers (#6747)
* Updating rails form hints to be labels

-- What
By updating the simple_form initializer wrappers and the configuration
of the output in the validated_form template, we ensure that the hints
-- if present -- are caught by screen readers. We do this by making
the hints a <label> instead of a <div>, and pointing that label to the
same input element as the main form's label. This way both the main
label and the hint are read by screen readers.

* Updating forgot password link preamble for screen reader

-- What
On the Re-Enter Password screen for GPO verify, the text "Forgot
password?" that precedes the password reset link is not readable by
screen readers. To deal with this, we add the localized text of
"Forgot password?" to the `title=` attribute of the link itself inside
of the corresponding React component. This ensures that screen readers
will read both the link text and the forgot password hint.

changelog: Improvements, Accessibility, updating password hint

* Switching to basic aria-label and updating tests

changelog: Improvements, Accessibility, updating aria-label forgot password

* Fixing linting errors in ruby template

changelog: Improvements, Linting, fixing template formatting

* Simplifying form hint accessibility with describedby

-- What
Instead of using two <label> elements to refer to the same input
field (which might cause issues and is in a grey area when it comes to
correcness), we simply give the hint div a generated id and prepend
that id to the input's `aria-describedby` attribute value.

(This is in the context of validated field components)

changelog: Improvements, Accessibility, simplifying validated field
hint accessibility

* Fixing formatting lints in erb file

changelog: Improvements, Formatting, fixing erb linting errors

* Changing forgot password link text per discussion

-- What
After some discussion regarding the accesibility/descriptive
properties of the "Forgot password?" aria label and subsequent "Follow
these instructions" link on the SSN confirmation password entry page
of IdV, we decided to change the link text and remove the non-link
description entirely.

This commit makes that change and updates the localizations (which
already existed elsewhere) to match other "forgot password" text
around the site. Additionally, we have updated some test assumptions
about querying.

* Updating corresponding template and localizations / tests

-- What
Because the FSMv2 was still enabled for the React component locally, I
was not seeing the template version. This commit updates the template
and also removes now unused localizations from the config.

I am also updating a couple of Ruby component tests that assumed a
single aria-describedby value on labelled validation components with
hints, whereas now we supply two values to that attribute

* Update spec/components/validated_field_component_spec.rb

Co-authored-by: Andrew Duthie <andrew.duthie@gsa.gov>

* Update app/javascript/packages/verify-flow/steps/password-confirm/password-confirm-step.tsx

Co-authored-by: Andrew Duthie <andrew.duthie@gsa.gov>

* Changing name of validate field hint element id

changelog: Improvements, Accessibility, element ids

* Switching field error aria description to come before field hint

-- What
Previously, the order of the aria-describedby attribute values was
reversed from what we would probably want, meaning that if a
validation error was present, the error would be read after the
hint. We want the error to be read first so users know that something
was not right, then provide the hint.

* Update spec/components/validated_field_component_spec.rb

Co-authored-by: Andrew Duthie <andrew.duthie@gsa.gov>