Updating check for mobile with special case for iPad
-- What
As of iOS 12 / the newest iPadOS versions, sites are by default
rendered with "full web" view mode, which causes the Safari on the
devices to report the incorrect user agent (in this case, a
Macintosh).

This prevents normal methods for checking whether or not the device is
an iPad.

As a proxy, we check for an apple user agent and then see whether or
not the device has `maxTouchPoints` of 5, which indicates an iPad.

changelog: Improvements, Device Compatibility, updating iPad checks