Updating upload steps mobile check to include frontend info
-- What
For some mobile devices -- iPad in particular -- we need explicit
information from the frontend about the client browser's capabilities
in order to properly determine if it is mobile and
camera-ready. However, on the Rails backend, we've been performing
server-side checking to determine whether or not a device is
mobile. This check is more or less a user agent check, which will
respond false for newer iPadOS devices.

In this commit, we add to our device_mobile? method to also check for
flow_session state that has been set from a hidden input on the
frontend. The input and session state setting were already occurring
-- this is not new code. What's new is checking for the presence of
the skip_upload_step property in the session, which only get set on
the frontend if the device is mobile.