Fixing email_sent_step_spec to force new method of mobile checking
-- What
Previously, this test relied on the Browser module's reading of the
User Agent, which was mocked to imitate an older mobile device. With
recent changes we are instead using the flow_session's
:skip_upload_step -- set indirectly by javascript on the frontend --
in the upload step's mobile_device? method. Here we simply stub that
method in this particular test suite.