Updating mailer image from svg to png for Google
-- What
Google proxies images that are sent in html emails that it receives in
any of its clients. However, Google does _not_ support SVG images in
html emails.

This caused us some problems when attempting to send a mailer with an
SVG image included, which worked in local testing (mailcatcher) but
failed in a broader integration test environment (INT), 404'ing on
Googles cached version of the image link.

This commit provides a plain png rendering of the original svg and
updates the relevant mailer template to display the png version. This
should get us past the Google problem.