Fixing broken tests
-- What
Mocha evidently has some difficulty changing the writability of
properties in before() functions when we are using the
useDefineProperty helper. This fixes that by just declaring the
properties we are concerned about as being writable at the top of the
file.

changelog: Improvements, Tablet Compatibility, updating frontend
platform check tests