LG-7277 Acuant Not Working for iPads (#6865)
* Updating check for mobile with special case for iPad

-- What
As of iOS 12 / the newest iPadOS versions, sites are by default
rendered with "full web" view mode, which causes the Safari on the
devices to report the incorrect user agent (in this case, a
Macintosh).

This prevents normal methods for checking whether or not the device is
an iPad.

As a proxy, we check for an apple user agent and then see whether or
not the device has `maxTouchPoints` of 5, which indicates an iPad.

changelog: Improvements, Tablet Compatibility, updating iPad checks

* Updating upload steps mobile check to include frontend info

-- What
For some mobile devices -- iPad in particular -- we need explicit
information from the frontend about the client browser's capabilities
in order to properly determine if it is mobile and
camera-ready. However, on the Rails backend, we've been performing
server-side checking to determine whether or not a device is
mobile. This check is more or less a user agent check, which will
respond false for newer iPadOS devices.

In this commit, we add to our device_mobile? method to also check for
flow_session state that has been set from a hidden input on the
frontend. The input and session state setting were already occurring
-- this is not new code. What's new is checking for the presence of
the skip_upload_step property in the session, which only get set on
the frontend if the device is mobile.

* Switching from "not" to bang "!"

changelog: Improvements, Tablet Compatibility, enabling Acuant capture
for ipads

* Fixing lint errors

changelog: Improvements, Tablet Compatibility, enabling acuant capture for ipad

* Simplifying mobile_device? check in upload step

changelog: Improvements, Tablet Compatibility, adding Acuant for iPad

* Fixing email_sent_step_spec to force new method of mobile checking

-- What
Previously, this test relied on the Browser module's reading of the
User Agent, which was mocked to imitate an older mobile device. With
recent changes we are instead using the flow_session's
:skip_upload_step -- set indirectly by javascript on the frontend --
in the upload step's mobile_device? method. Here we simply stub that
method in this particular test suite.

* Update app/javascript/packages/device/index.js

Co-authored-by: Andrew Duthie <andrew.duthie@gsa.gov>

* Updating tests to mock new session based mobile check

changelog: Improvements, Tablet Compatibility, enabling Acuant SDK for ipads

* Empty commit for CI

changelog: Improvements, Tablet Compatibility, adding iPad for Acuant SDK

* Fixing tests and lints

changelog: Improvements, Tablet Compatibility, assing Acuant for iPad

* Adding some signposting in upload step's params check

changelog: Improvements, Tablet Compatibility, adding Acuant for iPad

* Fixing spec lint

changelog: Improvements, Tablet Compatibility, adding Acuant for iPad

* Empty commit for CI

changelog: Improvements, Tablet Compatibility, adding Acuant for iPad

* Updating frontend mobile check spec

changelog: Improvements, Tablet Compatibility, Updating mobile checks on frontend

* Updating frontend mobile check tests

changelog: Improvements, Tablet Compatibility, updating tests for frontend
platform checks

* Fixing broken tests

-- What
Mocha evidently has some difficulty changing the writability of
properties in before() functions when we are using the
useDefineProperty helper. This fixes that by just declaring the
properties we are concerned about as being writable at the top of the
file.

changelog: Improvements, Tablet Compatibility, updating frontend
platform check tests

* Removing use of useDefineProperty in these tests

changelog: Improvements, Tablet Compatibility, updating platform check tests

Co-authored-by: Andrew Duthie <andrew.duthie@gsa.gov>