Updating FailedCaptureAtttempts context to track failed submissions
-- What
We need to track failed submissions in addition to failed captures,
and trigger forceNativeCamera accordingly.

The easiest way to accomplish this without touching the backend is to
simply add this as a property to the failed capture context, then
increment it on the review page.

In a subsequent commit, we should explore not even incrementing this,
but simply checking the upload remainingAttempts against the
configured maximum value, and setting forceNativeCamera based on that.

Updating tests accordingly.

changelog: Improvements, Acuant Capture, updating ability to force
native camera