Improve display of Process List in email notifications
**Why**: So that the Process List as displayed in email messages is consistent with the design system reference implementation.

It's assumed this is a side-effect of the revisions in #6736, where an increase in default email line-height of 1.3 to 1.5 caused an offset of the appearance of the text of the list item number. Explicitly assigning the line-height ensures correct math.

changelog: Bug Fixes, Email Templates, Improve appearance of Process List in email template