LG-7602 Pull VA IP Data and Store It In Session
https://cm-jira.usa.gov/browse/LG-7602

This is a supporting change for the LG-7602 feature for Inherited Proofing (IP)
where user PII needs to be stored in session for IP flow steps and
subsequent IDV workflow flow steps that need this information for (example)
Lexis Phone Finder look ups.

This change exposes a #user_pii method that retuns a Hash suitable to be
stored in the idv_session and flow_session[:pii_from_user] (minus :uuid).
The data mapped  to the user PII returned from this method is Service
Provider (SP)-specific in that the SP API returning the user PII may
be returned having data keys that differ from what we need; consequently,
create a mapper service for this, but that would disconnect what
the SP service returns hash-wise. The Form knows about this, so it
may be better suited "as is".

changelog: Upcoming Features, Inherited Proofing, Pull VA IP Data and Store It In Session (LG-7602)