https://cm-jira.usa.gov/browse/LG-7602
This is a supporting change for the LG-7602 feature for Inherited Proofing (IP).

The user's PII needs to be obtained from the Service Provider in our Inherited
Proofing process Flow Steps. Identifying whether or not an Inherited Proofing
SP request is in progress, needs access to the decorated_session, which currently
takes place in the InheritedProofingConcern. It makes sense (for how),
to serve up the Service Provider Services and Forms in this concern, since
we can easily identify the Service Provider here also; we can think about
placing this code in a SP Service and Form factory or something similar
later.

changelog: Upcoming Features, Inherited Proofing, Pull VA IP Data and Store It In Session (LG-7602)