Add logging of matching certificate for SAML AuthnRequests
Resolves LG-7676

Adds a new attribute to successful `SAML Auth` events to track the
serial number of the matching cert identified through signature
validation. If the signature validation fails, this attribute will be
`nil`, indicating that none of the configured SP public certificates
could be used to validate the signature in the SAML AuthnRequest.

changelog: Improvements, Authentication, Add logging of matching certificate for SAML AuthnRequests