Update long-running report queries to reset connection
**Why**: Occasionally we see PG::UnableToSend errors at which point
the connection stops receiving queries.

In an attempt to recover the job and continue, we call #reconnect!
on the connection. However, this would reset our session-specific
timeouts, so there's a big of rejiggering to allow the connection
timeout stuff from the reports to be called in to the db query
classes