LG-7891 Protect against nil state id numbers in mock proofer
Our dev docs have a sample YAML file which does not include a state ID number. As a result when partners change the jurisdiction to a supported jurisdiction it causes failures when calling downcase on the nil value.

This commit fixes this bug by protecting against a nil value there. We should also add a state id number and type to the example YAML on the dev docs.

[skip changelog]