Remove the selfie args from the `Idv::Agent`
The selfie is used to support strict IAL2. The strict IAL2 flow is being retired so we will not be matching selfies here in the near term. This commit retires the code that passed the selfie to the document proofing job and to the SDK adapter since it is not longer an active code path.