Remove the selfie args from the `Idv::Agent` (#7214)
The selfie is used to support strict IAL2. The strict IAL2 flow is being retired so we will not be matching selfies here in the near term. This commit retires the code that passes the selfie through the IdV agent to the document proofing job and to the SDK adapter since it is no longer an active code path.

[skip changelog]