Make sure that `selfie_image` and `liveness_checking_enabled` args aren't used and have default values in images clients
This commit gives us some safety and assurance that things won't break when we start deploying the changes to remove IAL2 strict.

All of the clients _should_ have a default arg for `selfie_image` and `liveness_checking_enabled` and none of the callers should be passing those args. We can follow this up with a commit to configure the clients to stop supporting those args.