Remove the last few selfie references in the error generator (#7241)
We are removing IAL2 strict and selfie match functionality. This commit cleans up the last references to the selfie match portions of the `ErrorGenerator`.

The `ErrorGenerator` produces a takes the reponse and ultimately generates an errors hash data structure which can be used to render errors to the user on specific images (e.g. the front, back, selfie, or entire ID). This commit removes the code that covered special cases for selfie failures.

[skip changelog]