Remove selfie and liveness enabled job and client args
The code that quits calling the document proofing job with the selfie image and liveness enabled args is merged and deployed to prod. As such it should be safe to remove these args form the job.

Additionally, the doc auth clients themselves should no longer be called with these args so they should be safe to remove from those as well.