Adding step-specific click continue method
-- What
Prior to this branch, all IDV step pages had the same "continue"
button, meaning `click_idv_continue` would find the same localization
on a button ("Continue" for english) and click that to proceed. The
new consolidated phone verification step breaks that expectation.

Instead of casing for this all over the test suite, we are introducing
a replacement `idv_click_continue_for_step` in the helpers, which
takes the symbol for the current step name as its argument and then
clicks the correct "continue" button -- which for the phone step is
now "Send security code" (and can be manually called from
`click_idv_send_security_code`).

This commit represents an initial implementation of this new helper
and a refactor of one of the example files that needs to use it.