Updating Send Link Step to send correct sms param for regular phone
-- What
The PhoneForm was changed to work with the otp phone verification, but
this interfered with other kinds of phone validation controls because
the form now expects an :otp_delivery_preference key. This being nil
will fail in other contexts.

With this commit we add a (rough, perhaps should be refactored) solution