Fixing sdkSrc and cameraSrc props 2b forced when present
-- What
We need the cameraSrc and sdkSrc props on the Acuant context, because
tests provide about:blank so that the loading of the scripts does not
throw errors (and success can be simulated on frontend).

This commit provides a mechanism by which these props are null by
default, but if there is a value present they will be used above any
other consideration. If null, we kick it to the ab test and then to
determine which path(s) to use.

changelog: Internal, AB Testing, Acuant SDK Upgrade AB Testing