Removing React Context for Acuant version A/B Test
-- What
As suggested in a thread by @aduth
[https://github.com/18F/identity-idp/pull/7392#discussion_r1033979065]
we probably don't need all the overhead of a new React Context for
dealing with the Acuant version a/b testing. Instead, we can surmise
what the camera and sdk routes should be from the AppRoot values and
simply pass these to the AcuantContextProvider.

This commit represents deletions, modifications, and test updates that
attempt to implement this suggestion.

As I see it, there are still two open questions that are not deal
breakers for finishing this work in the immediate term:
1. We declare the src routes as default prop values in the Acuant
context (React), but also specify these versions in the document
capture step file (ruby). Should we specify the new and old versions
as globals somewhere (IdentityConfig store)?
2. Do we have a trackEvent call for when the acuant SDK has loaded or
attempts to load? If not, should we add one? (I'm thinking about the
a/b test information being present at load time vs capture time vs
submission time)

changelog: Internal, AB Testing, Acuant SDK Upgrade Testing