Fix flaky GPO exporter test (#7400)
* Fix flaky GPO exporter test

Address a flaky test introduced in #7389.

We run tests in CI with a random Timezone each time. This means sometimes the local day and UTC day are different, which was making the GPO exporter test fail.

This change updates `GpoConfirmationExporter` to convert dates to UTC before outputting them. This gives us consistency on tests but should not matter in prod, where the Timezone is already set to UTC.

For LG-8233

[skip changelog]

* Update app/services/gpo_confirmation_exporter.rb

Co-authored-by: Zach Margolis <zachmargolis@users.noreply.github.com>

Co-authored-by: Zach Margolis <zachmargolis@users.noreply.github.com>