Scrub the error message from JSON parse errors of LexisNexis response bodies (#7412)
The LexisNexis response bodies may contain information that we do not wish to be logged or reported to NewRelic (such as PII).

When a JSON parse error is raised it includes the portion of the JSON that resulted in the error. This commit rescues the JSON parse errors that are raised when a LexisNexis response body is being parsed. These error messages may contain sensitive information. As a result a new JSON parse error is raised in their place with a message that does not include the JSON that caused the error.

[skip changelog]