LG-7393 Two-Screen Phone Verification (WIP) (#7330)
* change idv phone page to new text

* add otp selection to verfiy phone page

* added opt_delivery_preference to phone controller

* WIP fix tests in phone_controller_spec

* fix lint

* Setting SMS option as checked by default and adding test

* Adding new "send code" click helper

* Fixing re-entry test

* Updating failed phone job example with correct button

* Fixing some of the example test assumptions

* Adding step-specific click continue method

-- What
Prior to this branch, all IDV step pages had the same "continue"
button, meaning `click_idv_continue` would find the same localization
on a button ("Continue" for english) and click that to proceed. The
new consolidated phone verification step breaks that expectation.

Instead of casing for this all over the test suite, we are introducing
a replacement `idv_click_continue_for_step` in the helpers, which
takes the symbol for the current step name as its argument and then
clicks the correct "continue" button -- which for the phone step is
now "Send security code" (and can be manually called from
`click_idv_send_security_code`).

This commit represents an initial implementation of this new helper
and a refactor of one of the example files that needs to use it.

* Updating to correct path assumptions in phone step tests

* Adding missing method to phone controller

-- What
This method was previously on the otp_selection_controller, which is
on the way to being deprecated. Some existing code required that this
method now be present on the phone controller.

* Updating Send Link Step to send correct sms param for regular phone

-- What
The PhoneForm was changed to work with the otp phone verification, but
this interfered with other kinds of phone validation controls because
the form now expects an :otp_delivery_preference key. This being nil
will fail in other contexts.

With this commit we add a (rough, perhaps should be refactored) solution

* Updating helpers with new flow

-- What
Some of the helper methods still assumed an extra "select otp
verification" preference step. This commit removes those assumptions
from the helpers.

* Fixing tests

* Skipping failing test temporarily

-- What
This test -- unlike its counterpart on main -- will end up taking the
user to an account locked page, and for the moment it's not clear why.

* Fixing more tests, migrating methods

-- What
There were a couple of more methods from the deprecated
otp_selection_controller that needed to be moved to phone_controller.

There are also still some assumptions about the
otp_delivery_preference that are unclear

* Removing deprecated delivery selection step test

* Removing deprecated otp_selection_submitted from analytics

* Fixing deprecated calls to otp_delivery_selection_step helpers

* Removing old otp_delivery_method files

-- What
This is the step that has been combined with the phone number step,
and therefore it is no longer needed.

* Fixing phone controller spec assumptions

-- What
There were specific assumptions in place now about the default
delivery method needed on PhoneForm submissions.

* Removing old tests and fixing skipped tests

* Fixing lints

* Removing unused localizations

* Updating babel-loader to stable version

* Normalizing yaml

* Changing send_security_code to send_one_time_code in localization

* Add changelog

changelog: Features, Idv, consolidate phone verification to two screens

* fix changelog

changelog: Features, IdV, Consolidate phone verification to two screens

* added otp_delivery_preference to phone form analytics event

changelog: Improvements, IdV, Consolidate phone verification to two
screens

* change translation of security code to one-time-code

* remove test that rate limits otp_delivery_method_page

* remove unused translation

* Revert "Removing old otp_delivery_method files"

This reverts commit 7b26b8f1914985d5ac45cb1fa16128b0b97df35b.

* Revert "Removing deprecated otp_selection_submitted from analytics"

This reverts commit c252d796ee0e524843077758539c7f1fd81dfce9.

* Restoring missing localization keys for otp_delivery_method

[skip changelog]

* add back in missing translations

* check for vendor outage before sms

* allow redirect to old otp_delivery_method_page

* remove phone delivery otp from analytics regression tests

these events will soon be removed once the new two phone page gets
implemented. the otp_delivery_method selection step does not occur in
this regression test

* fix reporting of delivery method tracking to irs

* fix inputs to phone form in phone_controller_spec

* fix analytics on phone_controller

added in adapter to the analytics call for
idv_phone_confirmation_otp_sent

* phone otp resend test pass

* fix test

* move phone confirmation session to Idv::PhoneConfirmation session

* fixup

Co-authored-by: eric-gade <eric.gade@gsa.gov>
Co-authored-by: Eric Gade <105373963+eric-gade@users.noreply.github.com>